@@ -40,11 +40,11 @@ pub fn with_cxx_ffi_types_with_namespace(
     // builder = builder.with_sys_include("rust_types.h");
     for name in CXX_TYPE_NAMES {
         builder = builder.exclude_item(name);
-        builder = builder.rename_item(name, &format!("{}::{}", namespace, name).as_str());
+        builder = builder.rename_item(name, &format!("{namespace}::{name}").as_str());
     }
     for name in CXX_WRAPPER_NAMES {
         builder = builder.exclude_item(name);
-        builder = builder.rename_item(name, &format!("{}::{}", namespace, name).as_str());
+        builder = builder.rename_item(name, &format!("{namespace}::{name}").as_str());
     }
 
     builder
